refactors the modal queries
