+++ src/components/FamilyManifestModal/index.js
@@ -1,6 +1,2 @@
 export { default } from './FamilyManifestModal';
-export {
-  familyMemberAndParticipantDataQueryBody,
-  dataTypeDataQueryBody,
-  generateFamilyManifestModalProps,
-} from './FamilyManifestModalData';
+export { generateFamilyManifestModalProps } from './queries';
